+++ base/utils/R/do_conversions.R
@@ -66,11 +66,12 @@ do_conversions <- function(settings, overwrite.met = FALSE, overwrite.fia = FALS
       ## rather than having a model-format soils file that is processed remotely
     }
     # met conversion
+    
     if (input.tag == "met") {
       name <- ifelse(is.null(settings$browndog), "MET Process", "BrownDog")
       if ( (PEcAn.utils::status.check(name) == 0)) { ## previously is.null(input$path) && 
         PEcAn.logger::logger.info("calling met.process: ",settings$run$inputs[[i]][['path']])
-        settings$run$inputs[[i]][['path']] <- 
+        settings$run$inputs[[i]] <- 
           PEcAn.data.atmosphere::met.process(
             site       = settings$run$site, 
             input_met  = settings$run$inputs$met,