+++ models/ed/R/ed_varlookup.R
@@ -23,10 +23,10 @@ ed.var <- function(varname) {
                drelated = NULL,
                expr    = "DBH")                                                    
   } else if(varname == "AbvGrndWood") {
-    out = list(readvar = c("AGB_CO", "BLEAF"), 
+    out = list(readvar =  c("AGB_CO"), #until I change BLEAF keeper to be annual work with total AGB
                type = 'co', units = "kgC/plant", 
                drelated = NULL, 
-               expr    = "AGB_CO-BLEAF")                                           
+               expr    = "AGB_CO")                                           
   } else if(varname == "leaf_carbon_content") {
     out = list(readvar = "BLEAF", 
                type = 'co', units = "kgC/plant", 
@@ -38,10 +38,20 @@ ed.var <- function(varname) {
                drelated = NULL, 
                expr    = "BROOT")                                                  
   } else if(varname == "GWBI") {
-    out = list(readvar = "DLNAGB_DT", # this is actually rate of change in AGB 
-               type = 'co', units = "1/yr", 
+    out = list(readvar = "DDBH_DT", # this is actually rate of change in DBH, we'll calculate GWBI from it 
+               type = 'co', units = "cm/yr", 
                drelated = NULL,
-               expr    = "DLNAGB_DT")                                              
+               expr    = "DDBH_DT") 
+  } else if(varname == "fast_soil_pool_carbon_content") {
+    out = list(readvar = "FAST_SOIL_C", 
+               type = 'pa', units = "kg/m2", 
+               drelated = NULL,
+               expr    = "FAST_SOIL_C")    
+  } else if(varname == "structural_soil_pool_carbon_content") {
+    out = list(readvar = "STRUCTURAL_SOIL_C", 
+               type = 'pa', units = "kg/m2", 
+               drelated = NULL,
+               expr    = "STRUCTURAL_SOIL_C")    
   } else { # No Match!
     warning(paste0("Couldn't find varname ", varname, "!"))
     out = NULL