+++ modules/data.land/R/soil_process.R
@@ -11,23 +11,20 @@
 #' @examples
 soil_process <- function(settings, input, dbfiles, overwrite = FALSE,run.local=TRUE){
   
-  if(is.null(input$id)){
+  if(input$soil$source=="PalEON_soil" && is.null(input$id)){
     PEcAn.logger::logger.severe("currently soil_process requires an input ID to be specified")
     return(NULL)
   }
   
-  if(is.null(input$source)){
-    input$source <- "PalEON_soil"  ## temporarily hardcoding in the only source
+  if(is.null(input$soil$source)){
+    input$soil$source <- "PalEON_soil"  ## temporarily hardcoding in the only source
                                    ## in the future this should throw an error
   }
-  
-  #--------------------------------------------------------------------------------------------------#
   # Extract info from settings and setup
   site       <- settings$run$site
   model      <- settings$model$type
   host       <- settings$host
   dbparms    <- settings$database
-  
   # set up bety connection
   bety <- dplyr::src_postgres(dbname   = dbparms$bety$dbname, 
                               host     = dbparms$bety$host, 
@@ -35,7 +32,22 @@ soil_process <- function(settings, input, dbfiles, overwrite = FALSE,run.local=T
                               password = dbparms$bety$password)
   con <- bety$con
   on.exit(PEcAn.DB::db.close(con))
-  
+  # get site info
+  latlon <- PEcAn.data.atmosphere::db.site.lat.lon(site$id, con = con)
+  new.site <- data.frame(id = as.numeric(site$id), 
+                         lat = latlon$lat, 
+                         lon = latlon$lon)
+  str_ns <- paste0(new.site$id %/% 1e+09, "-", new.site$id %% 1e+09)
+  outfolder <- file.path(dbfiles, paste0(input$soil$source, "_site_", str_ns))
+  if(!dir.exists(outfolder)) dir.create(outfolder)
+  #--------------------------------------------------------------------------------------------------#   
+  # if we are reading from gSSURGO
+  if (input$soil$source=="gSSURGO"){
+    newfile<-extract_soil_gssurgo(outfolder,lat = latlon$lat,lon=latlon$lon)
+    return(newfile)
+  }
+  #--------------------------------------------------------------------------------------------------# 
+  # if we are reading  PalEON_soil
   # get existing input info
   source.input <- PEcAn.DB::db.query(paste0("SELECT * from Inputs where id =",input$id),con)
   if(run.local){
@@ -52,15 +64,8 @@ soil_process <- function(settings, input, dbfiles, overwrite = FALSE,run.local=T
       return(source.file)    
     }
   }  ## otherwise continue to process soil
-  
-  # get site info
-  latlon <- PEcAn.data.atmosphere::db.site.lat.lon(site$id, con = con)
-  new.site <- data.frame(id = as.numeric(site$id), 
-                         lat = latlon$lat, 
-                         lon = latlon$lon)
-  str_ns <- paste0(new.site$id %/% 1e+09, "-", new.site$id %% 1e+09)
-  
-  # set up host information
+
+    # set up host information
   machine.host <- ifelse(host == "localhost" || host$name == "localhost" || run.local,
                          PEcAn.remote::fqdn(), host$name)
   machine <- PEcAn.DB::db.query(paste0("SELECT * from machines where hostname = '", machine.host, "'"), con)
@@ -70,9 +75,6 @@ soil_process <- function(settings, input, dbfiles, overwrite = FALSE,run.local=T
     modeltype_id <- db.query(paste0("SELECT modeltype_id FROM models where id = '", settings$model$id, "'"), con)[[1]]
     model <- db.query(paste0("SELECT name FROM modeltypes where id = '", modeltype_id, "'"), con)[[1]]
   }
-  
-  outfolder <- file.path(dbfiles, paste0(input$source, "_site_", str_ns))
-  if(!dir.exists(outfolder)) dir.create(outfolder)
   
   newfile <- PEcAn.data.land::extract_soil_nc(source.file,outfolder,lat = latlon$lat,lon=latlon$lon)